refactor: chrome.storageへの依存をChromeStorageClientに限定する
